src/python/requirements.txt
PySide2==5.15.2
PyYAML==5.3.1
shiboken2==5.15.2
